Revert "apis updates"
